refactor(test): move all unit tests in a 'unit' folder
