+++ test/strategies/unit/node_wg.js
@@ -2,7 +2,7 @@
 import test from "tape";
 
 // Import Internal Dependencies
-import { SecurityWGStrategy } from "../../src/strategies/security-wg.js";
+import { SecurityWGStrategy } from "../../../src/strategies/security-wg.js";
 
 test("SecurityWGStrategy definition must return only two keys.", async(tape) => {
   const definition = await SecurityWGStrategy();
